// ssyweb/SSYNEW/src/app/models/LegacyClaims.ts
export class LegacyClaims {
    id: number;
    ssinNumber: string;
    oldRegdNo: string;
    claimRefernceNo: string;
    claimDate: any;
    claimDateString: string;
    claimAmount: any;
    status: number;
    statusName: string;
    claimCategory: number;
    claimCategoryName: string;
    remarks: string;
    userId: number;
    userType: number;
    //isInspector: boolean;
}